requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0
cloudinary>=1.40.0
insightface>=0.7.3
onnxruntime>=1.15.0
opencv-python>=4.8.0
scikit-image>=0.20.0